backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
google-generativeai